Move run command to top level
