+++ client_test/supports/app_run_tests/cli_args.py
@@ -1,0 +1,10 @@
+from datetime import datetime
+
+import modal
+
+stub = modal.Stub()
+
+
+@stub.local_entrypoint
+def func_with_args(dt: datetime):
+    print(f"the day is {dt.day}")
